cleanup: separate getDayEarnings to file
