+++ src/utils/getDayEarnings.ts
@@ -1,0 +1,11 @@
+const getDayEarnings = ({
+  appointmentDuration,
+  appointmentPrice,
+  workingHours,
+}: {
+  appointmentDuration: number;
+  appointmentPrice: number;
+  workingHours: number;
+}) => (60 / appointmentDuration) * appointmentPrice * workingHours;
+
+export default getDayEarnings;
